Extract outbox updater to own file
